+++ src/ray/core_worker/BUILD.bazel
@@ -22,7 +22,7 @@ ray_cc_library(
         ":future_resolver",
         ":generator_waiter",
         ":memory_store",
-        ":normal_task_submitter",
+        "//src/ray/core_worker/task_submission:normal_task_submitter",
         ":object_recovery_manager",
         ":plasma_store_provider",
         ":profile_event",
@@ -70,25 +70,6 @@ ray_cc_library(
         "//src/ray/common:task_common",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/util:process",
-    ],
-)
-
-ray_cc_library(
-    name = "core_worker_fiber",
-    hdrs = ["fiber.h"],
-    deps = [
-        "//src/ray/util:logging",
-        "@boost//:fiber",
-    ],
-)
-
-ray_cc_library(
-    name = "actor_submit_queue",
-    hdrs = ["transport/actor_submit_queue.h"],
-    deps = [
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "@com_google_absl//absl/types:optional",
     ],
 )
 
@@ -150,7 +131,7 @@ ray_cc_library(
     deps = [
         ":actor_creator",
         ":actor_handle",
-        ":actor_task_submitter",
+        "//src/ray/core_worker/task_submission:actor_task_submitter",
         ":common",
         ":core_worker_context",
         ":reference_count",
@@ -169,10 +150,10 @@ ray_cc_library(
     srcs = ["reference_count.cc"],
     hdrs = ["reference_count.h"],
     deps = [
-        ":lease_policy",
         "//src/ray/common:id",
         "//src/ray/protobuf:common_cc_proto",
         "//src/ray/pubsub:publisher",
+        "//src/ray/core_worker:lease_policy",
         "//src/ray/pubsub:subscriber",
         "//src/ray/rpc:core_worker_client",
         "//src/ray/rpc:grpc_server",
@@ -187,6 +168,7 @@ ray_cc_library(
     name = "lease_policy",
     srcs = ["lease_policy.cc"],
     hdrs = ["lease_policy.h"],
+    visibility = [":__subpackages__"],
     deps = [
         "//src/ray/common:id",
         "//src/ray/common:task_common",
@@ -215,29 +197,6 @@ ray_cc_library(
         "@boost//:circular_buffer",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/synchronization",
-        "@com_google_absl//absl/types:optional",
-    ],
-)
-
-ray_cc_library(
-    name = "out_of_order_actor_submit_queue",
-    srcs = ["transport/out_of_order_actor_submit_queue.cc"],
-    hdrs = ["transport/out_of_order_actor_submit_queue.h"],
-    deps = [
-        ":actor_submit_queue",
-        "//src/ray/common:id",
-        "@com_google_absl//absl/container:btree",
-        "@com_google_absl//absl/types:optional",
-    ],
-)
-
-ray_cc_library(
-    name = "sequential_actor_submit_queue",
-    srcs = ["transport/sequential_actor_submit_queue.cc"],
-    hdrs = ["transport/sequential_actor_submit_queue.h"],
-    deps = [
-        "actor_submit_queue",
-        "//src/ray/common:id",
         "@com_google_absl//absl/types:optional",
     ],
 )
@@ -295,44 +254,6 @@ ray_cc_library(
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/synchronization",
-    ],
-)
-
-ray_cc_library(
-    name = "dependency_resolver",
-    srcs = ["transport/dependency_resolver.cc"],
-    hdrs = ["transport/dependency_resolver.h"],
-    deps = [
-        ":actor_creator",
-        ":memory_store",
-        ":task_manager_interface",
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/container:flat_hash_set",
-    ],
-)
-
-ray_cc_library(
-    name = "actor_task_submitter",
-    srcs = ["transport/actor_task_submitter.cc"],
-    hdrs = ["transport/actor_task_submitter.h"],
-    deps = [
-        ":actor_creator",
-        ":actor_submit_queue",
-        ":core_worker_context",
-        ":dependency_resolver",
-        ":out_of_order_actor_submit_queue",
-        ":sequential_actor_submit_queue",
-        "//src/ray/common:asio",
-        "//src/ray/common:id",
-        "//src/ray/common:ray_object",
-        "//src/ray/gcs:gcs_pb_util",
-        "//src/ray/gcs/gcs_client:gcs_client_lib",
-        "//src/ray/rpc:core_worker_client",
-        "@com_google_absl//absl/base:core_headers",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/container:flat_hash_set",
     ],
 )
 
@@ -430,24 +351,5 @@ ray_cc_library(
         "//src/ray/protobuf:common_cc_proto",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/container:flat_hash_set",
-    ],
-)
-
-ray_cc_library(
-    name = "normal_task_submitter",
-    srcs = ["transport/normal_task_submitter.cc"],
-    hdrs = ["transport/normal_task_submitter.h"],
-    deps = [
-        ":actor_manager",
-        ":core_worker_context",
-        ":dependency_resolver",
-        ":lease_policy",
-        ":memory_store",
-        ":task_manager",
-        "//src/ray/common:id",
-        "//src/ray/gcs:gcs_pb_util",
-        "//src/ray/raylet_client:raylet_client_lib",
-        "//src/ray/rpc:core_worker_client",
-        "@com_google_absl//absl/base:core_headers",
     ],
 )
