+++ src/ray/core_worker/actor_manager.h
@@ -25,7 +25,7 @@
 #include "ray/core_worker/actor_creator.h"
 #include "ray/core_worker/actor_handle.h"
 #include "ray/core_worker/reference_count.h"
-#include "ray/core_worker/transport/actor_task_submitter.h"
+#include "ray/core_worker/task_submission/actor_task_submitter.h"
 #include "ray/gcs/gcs_client/gcs_client.h"
 namespace ray {
 namespace core {
