+++ src/ray/core_worker/task_submission/BUILD.bazel
@@ -1,0 +1,95 @@
+load("//bazel:ray.bzl", "ray_cc_library")
+
+ray_cc_library(
+    name = "dependency_resolver",
+    srcs = ["dependency_resolver.cc"],
+    hdrs = ["dependency_resolver.h"],
+    visibility = [":__subpackages__"],
+    deps = [
+        # XXX: remove.
+        "//src/ray/core_worker:memory_store",
+        "//src/ray/core_worker:actor_creator",
+        "//src/ray/core_worker:lease_policy",
+        "//src/ray/core_worker:task_manager_interface",
+        "//src/ray/common:id",
+        "//src/ray/common:task_common",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
+    ],
+)
+
+ray_cc_library(
+    name = "actor_submit_queue",
+    hdrs = ["actor_submit_queue.h"],
+    visibility = ["//visibility:private"],
+    deps = [
+        "//src/ray/common:id",
+        "//src/ray/common:task_common",
+        "@com_google_absl//absl/types:optional",
+    ],
+)
+
+ray_cc_library(
+    name = "out_of_order_actor_submit_queue",
+    srcs = ["out_of_order_actor_submit_queue.cc"],
+    hdrs = ["out_of_order_actor_submit_queue.h"],
+    visibility = [":__subpackages__"],
+    deps = [
+        ":actor_submit_queue",
+        "//src/ray/common:id",
+        "@com_google_absl//absl/container:btree",
+        "@com_google_absl//absl/types:optional",
+    ],
+)
+
+ray_cc_library(
+    name = "sequential_actor_submit_queue",
+    srcs = ["sequential_actor_submit_queue.cc"],
+    hdrs = ["sequential_actor_submit_queue.h"],
+    visibility = [":__subpackages__"],
+    deps = [
+        ":actor_submit_queue",
+        "//src/ray/common:id",
+        "@com_google_absl//absl/types:optional",
+    ],
+)
+
+ray_cc_library(
+    name = "actor_task_submitter",
+    srcs = ["actor_task_submitter.cc"],
+    hdrs = ["actor_task_submitter.h"],
+    visibility = ["//src/ray/core_worker:__pkg__", ":__subpackages__"],
+    deps = [
+        "//src/ray/core_worker:actor_creator",
+        ":actor_submit_queue",
+        ":dependency_resolver",
+        ":out_of_order_actor_submit_queue",
+        ":sequential_actor_submit_queue",
+        "//src/ray/common:asio",
+        "//src/ray/common:id",
+        "//src/ray/gcs:gcs_pb_util",
+        "//src/ray/rpc:core_worker_client",
+        "@com_google_absl//absl/base:core_headers",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
+    ],
+)
+
+ray_cc_library(
+    name = "normal_task_submitter",
+    srcs = ["normal_task_submitter.cc"],
+    hdrs = ["normal_task_submitter.h"],
+    visibility = ["//src/ray/core_worker:__pkg__", ":__subpackages__"],
+    deps = [
+        ":dependency_resolver",
+        "//src/ray/core_worker:lease_policy",
+        # XXX: remove
+        "//src/ray/core_worker:memory_store",
+        "//src/ray/core_worker:task_manager",
+        "//src/ray/common:id",
+        "//src/ray/gcs:gcs_pb_util",
+        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/rpc:core_worker_client",
+        "@com_google_absl//absl/base:core_headers",
+    ],
+)
