+++ src/ray/core_worker/task_submission/actor_task_submitter.h
@@ -14,8 +14,6 @@
 
 #pragma once
 
-#include <boost/asio/thread_pool.hpp>
-#include <boost/thread.hpp>
 #include <deque>
 #include <list>
 #include <memory>
@@ -31,15 +29,12 @@
 #include "absl/synchronization/mutex.h"
 #include "ray/common/asio/asio_util.h"
 #include "ray/common/id.h"
-#include "ray/common/ray_object.h"
 #include "ray/core_worker/actor_creator.h"
-#include "ray/core_worker/context.h"
 #include "ray/core_worker/store_provider/memory_store/memory_store.h"
-#include "ray/core_worker/transport/actor_submit_queue.h"
-#include "ray/core_worker/transport/dependency_resolver.h"
-#include "ray/core_worker/transport/out_of_order_actor_submit_queue.h"
-#include "ray/core_worker/transport/sequential_actor_submit_queue.h"
-#include "ray/gcs/gcs_client/gcs_client.h"
+#include "ray/core_worker/task_submission/actor_submit_queue.h"
+#include "ray/core_worker/task_submission/dependency_resolver.h"
+#include "ray/core_worker/task_submission/out_of_order_actor_submit_queue.h"
+#include "ray/core_worker/task_submission/sequential_actor_submit_queue.h"
 #include "ray/rpc/worker/core_worker_client.h"
 
 namespace ray {
