+++ src/ray/core_worker/task_submission/sequential_actor_submit_queue.h
@@ -20,35 +20,32 @@
 #include "absl/container/btree_map.h"
 #include "absl/types/optional.h"
 #include "ray/common/id.h"
-#include "ray/core_worker/transport/actor_submit_queue.h"
+#include "ray/core_worker/task_submission/actor_submit_queue.h"
 
 namespace ray {
 namespace core {
 
 /**
- * OutofOrderActorSubmitQueue sends request as soon as the dependencies are resolved.
- *
- * Under the hood, it keeps requests in two queues: pending queue and sending queue.
- * Emplaced request is inserted into pending queue; once the request's dependency
- * is resolved it's moved from pending queue into sending queue.
+ * SequentialActorSumitQueue extends IActorSubmitQueue and ensures tasks are send
+ * in the sequential order defined by the sequence no.
  */
-class OutofOrderActorSubmitQueue : public IActorSubmitQueue {
+class SequentialActorSubmitQueue : public IActorSubmitQueue {
  public:
-  explicit OutofOrderActorSubmitQueue(ActorID actor_id);
+  explicit SequentialActorSubmitQueue(ActorID actor_id);
   /// Add a task into the queue.
-  void Emplace(uint64_t position, const TaskSpecification &spec) override;
+  void Emplace(uint64_t sequence_no, const TaskSpecification &task_spec) override;
   /// If a task exists.
-  bool Contains(uint64_t position) const override;
+  bool Contains(uint64_t sequence_no) const override;
   /// If the task's dependencies were resolved.
-  bool DependenciesResolved(uint64_t position) const override;
+  bool DependenciesResolved(uint64_t sequence_no) const override;
   /// Mark a task's dependency resolution failed thus remove from the queue.
-  void MarkDependencyFailed(uint64_t position) override;
+  void MarkDependencyFailed(uint64_t sequence_no) override;
   /// Make a task's dependency is resolved thus ready to send.
-  void MarkDependencyResolved(uint64_t position) override;
+  void MarkDependencyResolved(uint64_t sequence_no) override;
   // Mark a task has been canceled.
   // If a task hasn't been sent yet, this API will guarantee a task won't be
   // popped via PopNextTaskToSend.
-  void MarkTaskCanceled(uint64_t position) override;
+  void MarkTaskCanceled(uint64_t sequence_no) override;
   /// Clear the queue and returns all tasks ids that haven't been sent yet.
   std::vector<TaskID> ClearAllTasks() override;
   /// Find next task to send.
@@ -60,9 +57,20 @@ class OutofOrderActorSubmitQueue : public IActorSubmitQueue {
   bool Empty() override;
 
  private:
-  ActorID kActorId;
-  absl::btree_map<uint64_t, std::pair<TaskSpecification, bool>> pending_queue_;
-  absl::btree_map<uint64_t, std::pair<TaskSpecification, bool>> sending_queue_;
+  /// The ID of the actor.
+  ActorID actor_id;
+
+  /// The actor's pending requests, ordered by the sequence number in the request.
+  /// The bool indicates whether the dependencies for that task have been resolved yet.
+  /// A task will be sent after its dependencies are resolved.
+  absl::btree_map<uint64_t, std::pair<TaskSpecification, bool>> requests;
+
+  /// Map of task retries. The bool indicates whether the dependencies for that task have
+  /// been resolved yet. A task will be sent after its dependencies are resolved. This is
+  /// a separate unordered map becuase the order in which retries are executed is
+  /// purposefully not guaranteed.
+  absl::flat_hash_map<uint64_t, std::pair<TaskSpecification, bool>> retry_requests;
 };
+
 }  // namespace core
 }  // namespace ray
