+++ src/ray/core_worker/task_submission/test/direct_actor_transport_mock_test.cc
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "ray/core_worker/transport/actor_task_submitter.h"
+#include "ray/core_worker/task_submission/actor_task_submitter.h"
 
 // clang-format off
 #include <memory>
