+++ src/ray/core_worker/test/BUILD.bazel
@@ -29,76 +29,6 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "actor_task_submitter_test",
-    srcs = ["actor_task_submitter_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_mock",
-        "//src/ray/common:asio",
-        "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:actor_creator",
-        "//src/ray/core_worker:reference_count",
-        "//src/ray/core_worker:task_manager",
-        "//src/ray/rpc:core_worker_client",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "direct_actor_transport_mock_test",
-    srcs = ["direct_actor_transport_mock_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_mock",
-        "//src/ray/core_worker:memory_store",
-        "//src/ray/core_worker:reference_count",
-        "//src/ray/core_worker:task_manager",
-        "//src/ray/gcs/gcs_client:gcs_client_lib",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "dependency_resolver_test",
-    size = "small",
-    srcs = ["dependency_resolver_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_mock",
-        "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:dependency_resolver",
-        "//src/ray/core_worker:memory_store",
-        "//src/ray/raylet_client:raylet_client_lib",
-        "//src/ray/rpc:core_worker_client",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "normal_task_submitter_test",
-    size = "small",
-    srcs = ["normal_task_submitter_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_fakes",
-        "//:ray_mock",
-        "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:core_worker_lib",
-        "//src/ray/core_worker:memory_store",
-        "//src/ray/raylet_client:raylet_client_lib",
-        "//src/ray/rpc:core_worker_client",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
     name = "reference_count_test",
     size = "small",
     srcs = ["reference_count_test.cc"],
@@ -127,24 +57,9 @@ ray_cc_test(
         "//src/ray/common:task_common",
         "//src/ray/common:test_util",
         "//src/ray/core_worker:memory_store",
-        "//src/ray/core_worker:normal_task_submitter",
         "//src/ray/core_worker:object_recovery_manager",
         "//src/ray/object_manager:object_manager_common",
         "//src/ray/raylet_client:raylet_client_lib",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "actor_submit_queue_test",
-    size = "small",
-    srcs = ["actor_submit_queue_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/common:asio",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:out_of_order_actor_submit_queue",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
